Return a data frame of cloc recognized languages and associated extensions

Some file extensions map to multiple languages:

cloc_recognized_languages()

Value

tibble

Details

.cl files could be Lisp or OpenCL

.d files could be D or dtrace

.f or .for files could be Fortran or Forth

.fs files could be Forth or F#

.inc files could be PHP or Pascal

.jl files could be Lisp or Julia

.m files could be MATLAB, Mathematica, Mercury, MUMPS, or Objective C

.pl files could be Perl or Prolog

.pp files could be Pascal or Puppet

.pro files could be IDL, Prolog, or a Qt Project

.ts files could be TypeScript or Qt Linguist

.v files Coq or Verilog/SystemVerilog

cloc has subroutines that attempt to identify the correct language based on the file's contents for these special cases. Language identification accuracy is a function of how much code the file contains; .m files with just one or two lines for example, seldom have enough information to correctly distinguish between MATLAB, Mercury, MUMPS, or Objective C.

Examples

cloc_recognized_languages()

#> # A tibble: 242 x 2
#>    lang           extensions            
#>    <chr>          <chr>                 
#>  1 ABAP           abap                  
#>  2 ActionScript   as                    
#>  3 Ada            ada, adb, ads, pad    
#>  4 ADSO/IDSM      adso                  
#>  5 Agda           agda, lagda           
#>  6 AMPLE          ample, dofile, startup
#>  7 Ant            build.xml, build.xml  
#>  8 ANTLR Grammar  g, g4                 
#>  9 Apex Trigger   trigger               
#> 10 Arduino Sketch ino, pde              
#> # … with 232 more rows